fix: handle emojis, HTML tags, markdown links and quotes properly
The dev.to anchor generation is a mess. By trying to match it,
I found many edge cases in GitLab and GitHub profiles as well.

This commit delegates the anchor generation to an interface,
with currently two different implementations: one for dev.to,
one for the rest (default).

Tests and specific code has been added to ensure the following
are properly handled:
* HTML tags,
* code (especially HTML) in backticks,
* markdown links,
* emojis,
* special quotes characters

There are still some edge cases, for example dev.to doesn't handle
duplicate anchors (that is, it doesn't suffix them with a number
to avoid clashes), but all in all most usual cases should now
work properly.