fix: support 4 backticks
Make it possible to use more than 3 backticks (or ~) for code markers.
This change also ensures it is possible to use a code block with 4 markers,
which itself contains code blocks with 3 markers.

Basically, it remembers what the start was, and considers the code
block end only when the same marker is found.