feat(web): improve options by showing inputs on the right
Make it easier to read by aligning the inputs at the right of the
labels.

Note that it required spans to be added after the input type checkbox,
because some browsers (Firefox) do not allow :before and :after
pseudo-elements on elements with no content (such as input).
The span also needs to be opened and closed (versus `<span />`), or else
safari doesn't close it at the right place.